fix drilldown rendring
